feat: HTTP 요청 메시지 읽는 부분 HttpRequest 로 이동
