+++ tomcat/src/main/java/org/apache/coyote/request/HttpRequest.java
@@ -2,6 +2,9 @@ package org.apache.coyote.request;
 
 import static org.apache.coyote.request.startline.HttpMethod.*;
 
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 import org.apache.coyote.cookie.Cookie;
@@ -12,12 +15,16 @@ import org.apache.coyote.request.startline.StartLine;
 
 public class HttpRequest {
 
+    private static final int KEY_INDEX = 0;
+    private static final int VALUE_INDEX = 1;
+    private static final String HEADER_DELIMITER = ": ";
     private static final String HTML_EXTENSION = ".html";
     private static final String QUERY_START_CHARACTER = "?";
     private static final String ROOT = "/";
     private static final String EXTENSION_CHARACTER = ".";
     private static final String DEFAULT_PAGE_URL = "/index.html";
     private static final String COOKIE_HEADER = "Cookie";
+    private static final String CONTENT_LENGTH = "Content-Length";
 
     private final StartLine startLine;
     private final HttpHeader headers;
@@ -38,13 +45,12 @@ public class HttpRequest {
                 HttpRequestBody.from(requestBody));
     }
 
-    public String getRequestUrlWithoutQuery() {
-        final String requestUrl = getRequestPath();
-        if (requestUrl.contains(QUERY_START_CHARACTER)) {
-            final int index = requestUrl.indexOf(QUERY_START_CHARACTER);
-            return requestUrl.substring(0, index);
-        }
-        return requestUrl;
+    public static HttpRequest readRequest(final BufferedReader bufferedReader) throws IOException {
+        final String httpStartLine = bufferedReader.readLine();
+        final Map<String, String> httpHeaderLines = readHttpHeaderLines(bufferedReader);
+        final String requestBody = readRequestBody(bufferedReader, httpHeaderLines);
+
+        return HttpRequest.of(httpStartLine, httpHeaderLines, requestBody);
     }
 
     public String getRequestPath() {
@@ -52,6 +58,54 @@ public class HttpRequest {
         requestUrl = makeDefaultRequestUrl(requestUrl);
 
         return requestUrl;
+    }
+
+    public HttpMethod getRequestMethod() {
+        return startLine.getMethod();
+    }
+
+    public Cookies getCookies() {
+        return cookies;
+    }
+
+    public Optional<Cookie> getJSessionCookie() {
+        return cookies.getJSessionCookie();
+    }
+
+    public QueryParams getQueryParams() {
+        if (startLine.getMethod().equals(GET)) {
+            return startLine.getQueryParams();
+        }
+        return requestBody.getBodyWithQueryParam();
+    }
+
+    private static Map<String, String> readHttpHeaderLines(BufferedReader bufferedReader) throws IOException {
+        final Map<String, String> httpHeaderLines = new HashMap<>();
+        String line;
+
+        while ((line = bufferedReader.readLine()) != null) {
+            if (line.isBlank()) {
+                break;
+            }
+            final String[] header = line.split(HEADER_DELIMITER);
+            httpHeaderLines.put(header[KEY_INDEX], header[VALUE_INDEX].trim());
+        }
+
+        return httpHeaderLines;
+    }
+
+    private static String readRequestBody(BufferedReader bufferedReader, Map<String, String> httpHeaderLines)
+            throws IOException {
+        final String contentLengthHeader = httpHeaderLines.get(CONTENT_LENGTH);
+        if (contentLengthHeader == null) {
+            return "";
+        }
+
+        final int contentLength = Integer.parseInt(contentLengthHeader.trim());
+        final char[] buffer = new char[contentLength];
+        bufferedReader.read(buffer, 0, contentLength);
+
+        return new String(buffer);
     }
 
     private String makeDefaultRequestUrl(String requestUrl) {
@@ -72,28 +126,5 @@ public class HttpRequest {
             return path + HTML_EXTENSION + QUERY_START_CHARACTER + queryString;
         }
         return requestUrl + HTML_EXTENSION;
-    }
-
-    public String getRequestBody() {
-        return requestBody.getRequestBody();
-    }
-
-    public HttpMethod getRequestMethod() {
-        return startLine.getMethod();
-    }
-
-    public Cookies getCookies() {
-        return cookies;
-    }
-
-    public Optional<Cookie> getJSessionCookie() {
-        return cookies.getJSessionCookie();
-    }
-
-    public QueryParams getQueryParams() {
-        if (startLine.getMethod().equals(GET)) {
-            return startLine.getQueryParams();
-        }
-        return requestBody.getBodyWithQueryParam();
     }
 }
